Separated profile creation from dashboard, started basic logic for handling first time profile creation
